updated sample app & base_model
